Modified the 1-print_binary.c file

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -1,14 +1,16 @@
 #include "main.h"
 
 /**
+ * print_binary - The function prints the binary representation of a number
+ * @n : The number to print its binary
  */
 void print_binary(unsigned long int n)
 {
 	unsigned long int i;
-	unsigned int msk = 1 << (sizeof(unsigned int) * 8 - 1);
+	unsigned long int msk = 1UL << (sizeof(unsigned long int) * 8 - 1);
 	bool leading_zeros = true;
 
-	for (i = 0; i < sizeof(unsigned int) * 8; i++)
+	for (i = 0; i < sizeof(unsigned long int) * 8; i++)
 	{
 		if (n & msk)
 		{
